third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc-expected.txt

PASS img.currentSrc is ""
PASS img.currentSrc is expectedCurrentSrc
PASS img.currentSrc is expectedCurrentSrc
PASS img_err.currentSrc is "http://127.0.0.1:8000/resources/slow-image.php?name=not_there.png&mimeType=image/png&sleep=300"
PASS img_redirect.currentSrc is "http://127.0.0.1:8000/resources/slow-image.php?redirect_name=square100.png&mimeType=image/png&sleep=500"
PASS successfullyParsed is true

TEST COMPLETE
